+++ common/model/src/main/java/is/codion/common/model/loadtest/LoadTest.java
@@ -61,9 +61,9 @@ public interface LoadTest<T> {
   Value<User> user();
 
   /**
-   * @return the load test title
+   * @return the load test name, or an empty Optional if none is available
    */
-  String title();
+  Optional<String> name();
 
   /**
    * Sets the random chooser weight for the given scenario
@@ -212,10 +212,10 @@ public interface LoadTest<T> {
     Builder<T> scenarios(Collection<? extends Scenario<T>> scenarios);
 
     /**
-     * @param titleFactory the title factory
+     * @param name the load test name
      * @return this builder
      */
-    Builder<T> titleFactory(Function<LoadTest<T>, String> titleFactory);
+    Builder<T> name(String name);
 
     /**
      * @return a new load test instance
@@ -262,6 +262,8 @@ public interface LoadTest<T> {
   /**
    * Specifies a load test usage scenario.
    * @param <T> the type used to run the scenario
+   * @see #scenario(Performer)
+   * @see #scenario(Performer, int)
    * @see #builder(Performer)
    */
   interface Scenario<T> {
@@ -340,6 +342,27 @@ public interface LoadTest<T> {
      */
     static <T> Builder<T> builder(Performer<T> performer) {
       return new DefaultScenario.DefaultBuilder<>(performer);
+    }
+
+    /**
+     * @param performer the scenario performer
+     * @return a new {@link Scenario} instance
+     * @param <T> the load test application type
+     */
+    static <T> Scenario<T> scenario(Performer<T> performer) {
+      return builder(performer).build();
+    }
+
+    /**
+     * @param performer the scenario performer
+     * @param defaultWeight the default scenario weight
+     * @return a new {@link Scenario} instance
+     * @param <T> the load test application type
+     */
+    static <T> Scenario<T> scenario(Performer<T> performer, int defaultWeight) {
+      return builder(performer)
+              .defaultWeight(defaultWeight)
+              .build();
     }
 
     /**