+++ demos/petstore/src/main/java/is/codion/framework/demos/petstore/testing/PetstoreLoadTest.java
@@ -25,12 +25,12 @@ import is.codion.common.user.User;
 import is.codion.framework.db.EntityConnectionProvider;
 import is.codion.framework.demos.petstore.domain.Petstore;
 import is.codion.framework.demos.petstore.model.PetstoreAppModel;
-import is.codion.swing.common.model.tools.loadtest.LoadTestModel;
-import is.codion.swing.common.ui.tools.loadtest.LoadTestPanel;
 import is.codion.swing.framework.model.SwingEntityModel;
 
 import java.util.function.Function;
 
+import static is.codion.swing.common.model.tools.loadtest.LoadTestModel.loadTestModel;
+import static is.codion.swing.common.ui.tools.loadtest.LoadTestPanel.loadTestPanel;
 import static is.codion.swing.framework.model.tools.loadtest.EntityLoadTestUtil.selectRandomRow;
 import static java.util.Collections.singletonList;
 
@@ -83,8 +83,8 @@ public final class PetstoreLoadTest {
                     .scenarios(singletonList(Scenario.builder(new PetstoreUsageScenario())
                             .name("selectRecords")
                             .build()))
-                    .titleFactory(model -> "Petstore LoadTest - " + EntityConnectionProvider.CLIENT_CONNECTION_TYPE.get())
+                    .name("Petstore LoadTest - " + EntityConnectionProvider.CLIENT_CONNECTION_TYPE.get())
                     .build();
-    new LoadTestPanel<>(LoadTestModel.loadTestModel(loadTest)).run();
+    loadTestPanel(loadTestModel(loadTest)).run();
   }
 }