+++ swing/common-ui-tools/src/main/java/is/codion/swing/common/ui/tools/loadtest/LoadTestPanel.java
@@ -81,6 +81,7 @@ import static org.jfree.chart.ChartFactory.createXYStepChart;
 /**
  * A default UI component for the LoadTestModel class.
  * @param <T> the load test application type
+ * @see #loadTestPanel(LoadTestModel)
  * @see LoadTestModel
  */
 public final class LoadTestPanel<T> extends JPanel {
@@ -92,6 +93,7 @@ public final class LoadTestPanel<T> extends JPanel {
   private static final int SPINNER_STEP_SIZE = 10;
   private static final double RESIZE_WEIGHT = 0.8;
   private static final NumberFormat DURATION_FORMAT = NumberFormat.getIntegerInstance();
+  private static final String DEFAULT_TITLE = "Codion LoadTest";
 
   private final LoadTestModel<T> loadTestModel;
   private final LoadTest<T> loadTest;
@@ -107,11 +109,7 @@ public final class LoadTestPanel<T> extends JPanel {
 
   private boolean exiting;
 
-  /**
-   * Constructs a new LoadTestPanel.
-   * @param loadTestModel the LoadTestModel to base this panel on
-   */
-  public LoadTestPanel(LoadTestModel<T> loadTestModel) {
+  private LoadTestPanel(LoadTestModel<T> loadTestModel) {
     this.loadTestModel = requireNonNull(loadTestModel, "loadTestModel");
     this.loadTest = loadTestModel.loadTest();
     this.loadTestModel.applicationTableModel().refresher().addRefreshFailedListener(this::displayException);
@@ -133,6 +131,16 @@ public final class LoadTestPanel<T> extends JPanel {
   }
 
   /**
+   * Instantiates a new {@link LoadTestPanel} instance.
+   * @param loadTestModel the LoadTestModel to base this panel on
+   * @return a new {@link LoadTestPanel} instance.
+   * @param <T> the load test application type
+   */
+  public static <T> LoadTestPanel<T> loadTestPanel(LoadTestModel<T> loadTestModel) {
+    return new LoadTestPanel<>(loadTestModel);
+  }
+
+  /**
    * Shows a frame containing this load test panel
    * @return the frame
    */
@@ -140,7 +148,7 @@ public final class LoadTestPanel<T> extends JPanel {
     return frame(this)
             .icon(logoTransparent())
             .menuBar(menu(createMainMenuControls()).createMenuBar())
-            .title("Codion - " + loadTest.title())
+            .title(loadTest.name().orElse(DEFAULT_TITLE))
             .defaultCloseOperation(WindowConstants.DISPOSE_ON_CLOSE)
             .onClosing(windowEvent -> exit())
             .size(screenSizeRatio(DEFAULT_SCREEN_SIZE_RATIO))