+++ swing/common-ui-tools/src/test/java/is/codion/swing/common/ui/tools/loadtest/LoadTestPanelTest.java
@@ -24,6 +24,8 @@ import is.codion.swing.common.model.tools.loadtest.LoadTestModel;
 
 import org.junit.jupiter.api.Test;
 
+import static is.codion.swing.common.model.tools.loadtest.LoadTestModel.loadTestModel;
+import static is.codion.swing.common.ui.tools.loadtest.LoadTestPanel.loadTestPanel;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
@@ -38,14 +40,14 @@ public class LoadTestPanelTest {
             .loginDelayFactor(2)
             .applicationBatchSize(2)
             .build();
-    LoadTestModel<Object> model = LoadTestModel.loadTestModel(loadTest);
-    LoadTestPanel<Object> panel = new LoadTestPanel<>(model);
+    LoadTestModel<Object> model = loadTestModel(loadTest);
+    LoadTestPanel<Object> panel = loadTestPanel(model);
     assertEquals(model, panel.model());
     loadTest.shutdown();
   }
 
   @Test
   void constructorNullModel() {
-    assertThrows(NullPointerException.class, () -> new LoadTestPanel<LoadTest<?>>(null));
+    assertThrows(NullPointerException.class, () -> loadTestPanel(null));
   }
 }